+++ src/dHedgePPM_1.jl
@@ -2,15 +2,15 @@
 
 # With p-loss
 mutable struct dHedgePPM_1{T} <: BasePredictor{T}
-    model::Trie{T,Int64}
+    model::Trie{T,Int}
     context::Vector{T}
-    cxt_length::Int64
+    cxt_length::Int
     β::Float64
     γ::Float64
     weights::Vector{Float64}
     last_prediction::Vector{Dict{T,Float64}}
 
-    dHedgePPM_1{T}( _c::Int, _β::Float64 = 1.0, _γ::Float64 = 1.0 ) where {T} = new( Trie{T,Int64}(), Vector{T}(),
+    dHedgePPM_1{T}( _c::Int, _β::Float64 = 1.0, _γ::Float64 = 1.0 ) where {T} = new( Trie{T,Int}(), Vector{T}(),
                                                 _c, _β, _γ,
                                                 ones(Float64,_c+1), fill(Dict{T,Float64}(),_c+1) );
 end
@@ -93,7 +93,7 @@ end
 
 function unique_string( p::dHedgePPM_1{T} ) where {T}
     return @sprintf( "dHedgePPM1_%02d_%03d_%03d", p.cxt_length,
-                        trunc(Int64,100*p.β), trunc(Int64,100*p.γ)  );
+                        trunc(Int,100*p.β), trunc(Int,100*p.γ)  );
 end
 
 function predict_from_subcontext( p::dHedgePPM_1{T}, sub_cxt::Vector{T} )  where {T}
