+++ test/KOM.jl
@@ -17,7 +17,7 @@ data    = [ 'a', 'b', 'c', 'c', 'd',
 end
 
 # Test model
-test_model = Dict{Vector{Char},Int64}(
+test_model = Dict{Vector{Char},Int}(
     ['a'] => 1,
     ['a','b'] => 1,
     ['a','b','c'] => 1,
@@ -76,7 +76,7 @@ end
     # Reusing p
     @test get_best_symbol( p ) == 'b'
     # For untrained model
-    @test get_best_symbol( KOM{Int64}(10) ) == nothing
+    @test get_best_symbol( KOM{Int}(10) ) == nothing
     @test get_best_symbol( KOM{Char}(10) ) == nothing
     @test get_best_symbol( KOM{String}(10) ) == nothing
 end
@@ -88,9 +88,9 @@ end
 
 @testset "Random Sequence test" begin
     for idx = 1:10
-        learnt_model    = KOM{Int64}(5);
+        learnt_model    = KOM{Int}(5);
         @test @test_nothrow for i = 1:1000
-            symbol = trunc(Int64,10*rand());
+            symbol = trunc(Int,10*rand());
             add!( learnt_model, symbol );
             predict( learnt_model );
         end
