+++ test/ewPPM.jl
@@ -18,7 +18,7 @@ data    = [ 'a', 'b', 'c', 'c', 'd',
 end
 
 # Test model
-test_model = Dict{Vector{Char},Int64}(
+test_model = Dict{Vector{Char},Int}(
     ['a'] => 1,
     ['a','b'] => 1,
     ['a','b','c'] => 1,
@@ -77,7 +77,7 @@ end
     # Reusing p
     # @test get_best_symbol( p ) == 'b'
     # For untrained model
-    @test get_best_symbol( ewPPM{Int64}(4) ) == nothing
+    @test get_best_symbol( ewPPM{Int}(4) ) == nothing
     @test get_best_symbol( ewPPM{Char}(4) ) == nothing
     @test get_best_symbol( ewPPM{String}(4) ) == nothing
 end
@@ -89,9 +89,9 @@ end
 
 @testset "Random Sequence test" begin
     for idx = 1:10
-        learnt_model    = ewPPM{Int64}(5,0.75);
+        learnt_model    = ewPPM{Int}(5,0.75);
         @test @test_nothrow for i = 1:1000
-            symbol = trunc(Int64,10*rand());
+            symbol = trunc(Int,10*rand());
             add!( learnt_model, symbol );
             predict( learnt_model );
         end
